refactor : 소켓 연결 훅으로 분리
